rouge table-line code highlighting  and debugging: https://github.com/metanorma/metanorma-ogc/issues/465
