+++ lib/isodoc/base_style/reset.scss
@@ -153,10 +153,37 @@ a.FootnoteRef + a.FootnoteRef:before {
   text-decoration: line-through;
 }
 
-.rouge-gutter {
+/* code highlighting with line numbers */
+
+table.rouge-line-table td.rouge-gutter {
+  -moz-user-select: none;
+  -ms-user-select: none;
+  -webkit-user-select: none;
   user-select: none;
+  padding-right: 1em;
 }
 
-.rouge-code {
+table.rouge-line-table td.rouge-code {
+  -moz-user-select: all;
+  -ms-user-select: all;
+  -webkit-user-select: all;
   user-select: all;
 }
+
+table.rouge-line-table,
+table.rouge-line-table th,
+table.rouge-line-table td {
+  width: auto;
+  border: none;
+  margin: 0;
+  padding: 0;
+  font-size: 100%;
+}
+
+table.rouge-line-table pre {
+  margin: 0;
+  padding: 0;
+  overflow-x: visible;
+  font-size: 100%;
+}
+
