+++ lib/isodoc/base_style/typography.scss
@@ -65,6 +65,6 @@
     padding: 1em;
   }
   td.header {
-    font-weigth: 400;
+    font-weight: 400;
   }
 }
