+++ lib/isodoc/function/to_word_html.rb
@@ -8,10 +8,6 @@ module IsoDoc
         @termdomain = termdomain
       end
 
-      def in_sourcecode
-        @sourcecode
-      end
-
       def note?
         @note
       end
